TASK: Remove dependency to TYPO3SniffPool
As it's not managed anymore, we copy all necessary sniffs into this
repository. We therefore do not need to include sniffs in ruleset.xml
anymore, as they are auto included in this standard.

Furthermore we adjusted the sniffs to follow our CGL.

Also we removed some small parts which didn't work, or were excluded
anyway.

As comments were not in sync with code, we also removed them.